# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for demonstration purposes.",
        "title": "My API",
        "contact": {},
        "version": "1.0"
    },
    "host": "8.218.102.112:4880",
    "basePath": "/api/v1",
    "paths": {
...
        "/token_info/{chainType}/{tokenAddress}": {
            "get": {
                "description": "Retrieve detailed information for a specific token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Get token information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Type of the token (e.g., sol, eth)",
                        "name": "chainType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Address of the token",
                        "name": "tokenAddress",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detailed token information",
                        "schema": {
                            "$ref": "#/definitions/response.TokenInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.TokenInfoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/token_launchpad_info/{chainType}/{tokenAddress}": {
            "get": {
                "description": "Retrieve detailed information for a specific token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Get token launchpad info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Type of the token (e.g., sol, eth)",
                        "name": "chainType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Address of the token",
                        "name": "tokenAddress",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.TokenLaunchpadInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/token_market_analytics/{chainType}/{tokenAddress}": {
            "get": {
                "description": "Retrieve market analytics for a specific token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Get token market analytics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Type of the token (e.g., sol, eth)",
                        "name": "chainType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Address of the token",
                        "name": "tokenAddress",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.TokenMarketAnalyticsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/token_order_book/{chainType}/{tokenAddress}": {
            "get": {
                "description": "Retrieve order book information for a specific token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Get token order book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Type of the token (e.g., sol, eth)",
                        "name": "chainType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Address of the token",
                        "name": "tokenAddress",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.TokenOrderBookItem"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/token_prices/{chainType}": {
            "post": {
                "description": "Retrieve prices for a list of tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "Get token prices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Type of the token (e.g., sol, eth)",
                        "name": "chainType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Array of token addresses",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.TokenPriceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/response.TokenPriceResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/tokens/{chainType}/{tokenAddress}/search": {
            "get": {
                "description": "Search for tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Search for tokens",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Query",
                        "name": "query",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.TokenPriceRequest": {
            "type": "object",
            "required": [
                "addresses"
            ],
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "response.TokenInfoResponse": {
            "description": "Detailed information about a token",
            "type": "object",
            "properties": {
                "address": {
                    "description": "The unique address of the token",
                    "type": "string"
                },
                "biggest_pool_address": {
                    "description": "Address of the largest liquidity pool for the token",
                    "type": "string"
                },
                "circulating_supply": {
                    "description": "The amount of tokens currently in circulation",
                    "type": "integer"
                },
                "create_timestamp": {
                    "description": "The timestamp when the token was created",
                    "type": "integer"
                },
                "created_platform_type": {
                    "description": "The platform type where the token was created",
                    "type": "integer"
                },
                "creator": {
                    "description": "The address of the token's creator",
                    "type": "string"
                },
                "creator_balance_rate": {
                    "description": "The percentage of total supply held by the creator",
                    "type": "string"
                },
                "creator_token_balance": {
                    "description": "The balance of tokens held by the creator",
                    "type": "string"
                },
                "cto_flag": {
                    "description": "Security check data",
                    "type": "boolean"
                },
                "decimals": {
                    "description": "The number of decimal places the token uses",
                    "type": "integer"
                },
                "dev_native_token_amount": {
                    "description": "The amount of native tokens held by the developer",
                    "type": "integer"
                },
                "dev_token_burn_amount": {
                    "description": "The amount of tokens burned by the developer",
                    "type": "number"
                },
                "dev_token_burn_ratio": {
                    "description": "The ratio of tokens burned by the developer",
                    "type": "number"
                },
                "dexscr_ad": {
                    "description": "Advertisement status on DEX",
                    "type": "boolean"
                },
                "dexscr_update_link": {
                    "description": "Update link status on DEX",
                    "type": "boolean"
                },
                "dexscr_update_link_time": {
                    "description": "Timestamp of the last update link on DEX",
                    "type": "integer"
                },
                "freeze_authority": {
                    "description": "Indicates if the token has freezing authority",
                    "type": "boolean"
                },
                "holder_count": {
                    "description": "The number of holders of the token",
                    "type": "integer"
                },
                "is_complete": {
                    "description": "Flag indicating if the token is a common token",
                    "type": "boolean"
                },
                "liquidity": {
                    "description": "The liquidity of the token in the market",
                    "type": "number"
                },
                "logo": {
                    "description": "URL to the token's logo image",
                    "type": "string"
                },
                "market_cap": {
                    "type": "number"
                },
                "max_supply": {
                    "description": "The maximum supply of the token",
                    "type": "integer"
                },
                "mint_authority": {
                    "description": "Indicates if the token has minting authority",
                    "type": "boolean"
                },
                "name": {
                    "description": "The name of the token",
                    "type": "string"
                },
                "open_timestamp": {
                    "description": "The timestamp when the token was first opened for trading",
                    "type": "integer"
                },
                "price": {
                    "description": "The price of the token",
                    "type": "number"
                },
                "price_change": {
                    "description": "The price change of the token",
                    "type": "number"
                },
                "progress": {
                    "description": "The progress of the token",
                    "type": "number"
                },
                "rat_trader_amount_rate": {
                    "description": "The percentage of total supply held by rat traders",
                    "type": "number"
                },
                "symbol": {
                    "description": "The symbol representing the token",
                    "type": "string"
                },
                "telegram": {
                    "description": "The Telegram handle of the token",
                    "type": "string"
                },
                "token_burn_ratio": {
                    "description": "The ratio of tokens burned by the token",
                    "type": "number"
                },
                "top10_holdings": {
                    "description": "The percentage of total supply held by the top 10 holders",
                    "type": "number"
                },
                "top_10_holder_rate": {
                    "description": "The percentage of total supply held by the top 10 holders",
                    "type": "number"
                },
                "total_holders": {
                    "description": "The total number of holders of the token",
                    "type": "boolean"
                },
                "total_supply": {
                    "description": "The total supply of the token",
                    "type": "integer"
                },
                "twitter": {
                    "description": "The Twitter handle of the token",
                    "type": "string"
                },
                "twitter_change_flag": {
                    "description": "Twitter change flag",
                    "type": "boolean"
                },
                "website": {
                    "description": "The website of the token",
                    "type": "string"
                }
            }
        },
        "response.TokenLaunchpadInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "launchpad": {
                    "type": "string"
                },
                "launchpad_progress": {
                    "type": "string"
                },
                "launchpad_status": {
                    "type": "integer"
                }
            }
        },
        "response.TokenMarketAnalyticsResponse": {
            "type": "object",
            "properties": {
                "buy_count_1h": {
                    "type": "integer"
                },
                "buy_count_1m": {
                    "type": "integer"
                },
                "buy_count_24h": {
                    "type": "integer"
                },
                "buy_count_5m": {
                    "type": "integer"
                },
                "buy_volume_1h": {
                    "type": "number"
                },
                "buy_volume_1m": {
                    "type": "number"
                },
                "buy_volume_24h": {
                    "type": "number"
                },
                "buy_volume_5m": {
                    "type": "number"
                },
                "current_price": {
                    "type": "number"
                },
                "price_change_1h": {
                    "type": "number"
                },
                "price_change_1m": {
                    "type": "number"
                },
                "price_change_24h": {
                    "type": "number"
                },
                "price_change_5m": {
                    "type": "number"
                },
                "sell_count_1h": {
                    "type": "integer"
                },
                "sell_count_1m": {
                    "type": "integer"
                },
                "sell_count_24h": {
                    "type": "integer"
                },
                "sell_count_5m": {
                    "type": "integer"
                },
                "sell_volume_1h": {
                    "type": "number"
                },
                "sell_volume_1m": {
                    "type": "number"
                },
                "sell_volume_24h": {
                    "type": "number"
                },
                "sell_volume_5m": {
                    "type": "number"
                },
                "token_address": {
                    "type": "string"
                },
                "total_count_1h": {
                    "type": "integer"
                },
                "total_count_1m": {
                    "type": "integer"
                },
                "total_count_24h": {
                    "type": "integer"
                },
                "total_count_5m": {
                    "type": "integer"
                },
                "volume_1h": {
                    "type": "number"
                },
                "volume_1m": {
                    "type": "number"
                },
                "volume_24h": {
                    "type": "number"
                },
                "volume_5m": {
                    "type": "number"
                }
            }
        },
        "response.TokenOrderBookItem": {
            "description": "Single transaction record in the token order book",
            "type": "object",
            "properties": {
                "base_token_amount": {
                    "description": "The amount of base tokens in the transaction",
                    "type": "number"
                },
                "base_token_price": {
                    "description": "The price of the base token",
                    "type": "number"
                },
                "chain_type": {
                    "description": "The type of blockchain network",
                    "type": "integer"
                },
                "platform_type": {
                    "description": "The type of platform where the transaction occurred",
                    "type": "integer"
                },
                "pool_address": {
                    "description": "The address of the liquidity pool",
                    "type": "string"
                },
                "quote_token_amount": {
                    "description": "The amount of quote tokens in the transaction",
                    "type": "number"
                },
                "quote_token_price": {
                    "description": "The price of the quote token",
                    "type": "number"
                },
                "token_address": {
                    "description": "The address of the token being traded",
                    "type": "string"
                },
                "transaction_hash": {
                    "description": "The unique hash of the transaction",
                    "type": "string"
                },
                "transaction_time": {
                    "description": "The timestamp of the transaction in unix seconds",
                    "type": "integer"
                },
                "transaction_type": {
                    "description": "The type of transaction (buy/sell)",
                    "type": "integer"
                },
                "usd_amount": {
                    "description": "The transaction amount in USD",
                    "type": "number"
                },
                "user_address": {
                    "description": "The address of the user who made the transaction",
                    "type": "string"
                }
            }
        },
        "response.TokenPriceResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
    }
}